rename all to InAppMessages (plural)
