--- packages/react-core/src/modules/in-app-message/context/index.ts
@@ -1,1 +1,0 @@
-export * from "./KnockInAppChannelProvider";
