+++ packages/react-core/src/modules/in-app-messages/context/index.ts
@@ -1,0 +1,1 @@
+export * from "./KnockInAppMessagesChannelProvider";
