+++ packages/react/src/index.ts
@@ -2,6 +2,7 @@ import "./theme.css";
 
 export * from "./modules/core";
 export * from "./modules/feed";
-export * from "./modules/in-app-message";
+export * from "./modules/in-app-messages";
 export * from "./modules/slack";
+
 export * from "@knocklabs/react-core";
